requirements.txt
# Bird Feeding Project Dependencies
flask>=2.3.0
requests>=2.28.0

# Observability and logging
structlog>=22.0.0
python-json-logger>=2.0.0

# Java integration
JPype1>=1.4.0

# Additional dependencies (uncomment as needed):
# numpy>=1.21.0
# pandas>=1.5.0
